Consider a "stopped" Xplenty job as a failure
When an Xplenty job is "Stopped", it's unlikely it finished doing what it was meant to do. Considering such a job as successful leads to errors in DAGs due to expected files / tables / data not being present. Consider such Xplenty jobs as failures and raise an exception to fail the sensor.